Updated Geolocation Services
Geolocation services have now been moved from the JS of each Twine game to the Kotlin backend of the Application. User coordinates are written to a variable every 5 seconds, and the most recent coordinates can be stored in a JSONObject upon request via the JavaScriptInterface and are provided to the game. Users seeking to develop their own games that utilize Geolocation services now don't need to build out any of the Geolocation systems themselves, and can instead just request the JSON object via their story javascript.