Rebuilt Location Service logic
Discovered that the location service was not actually running in the background when using WebViewActivity, it only ran when MainActivity was focused. Rebuilt the logic to ensure that the LocationService runs in the foreground.

This likely solves the earlier "Variable Update" bug. Since the location service was not correctly running in the foreground, the user location was not updating. This meant that the same value was consistently being used in the hexagon checks, making it seem like the user was standing still when they were moving. The new location service provides updates even if MainActivity is not running, meaning that there is a consistent stream of new location data being provided to the WebView